Integer type converted to Long

+++ sdk/src/main/java/cryptohopper/android/sdk/API/Hopper/Position/UpdatePosition/HopperAPIUpdatePositionRequest.kt
@@ -10,7 +10,7 @@ import cryptohopper.android.sdk.SharedModels.ConfigModels.HopperAPIHttpMethod
 
 class HopperAPIUpdatePositionRequest: HopperAPIRequest<HopperCommonMessageResponse> {
     
-    constructor(hopperId : String , positionId:Int,takeProfit : Double,stopLoss:Int,stopLossPercentage : Double,trailingStopLoss : Int, trailingStopLossPercentage : Double,trailingStopLossArm : Double,autoClose : Int,autoCloseTime : Int) {
+    constructor(hopperId : String , positionId:Int,takeProfit : Double,stopLoss:Int,stopLossPercentage : Double,trailingStopLoss : Int, trailingStopLossPercentage : Double,trailingStopLossArm : Double,autoClose : Int,autoCloseTime : Long) {
         this.httpMethod = HopperAPIHttpMethod.POST
         this.needsAuthentication = true
         this.changeUrlPath("/hopper/$hopperId/position/setpositionsetting")
